requirements.txt
Django==1.9.2
argparse==1.2.1
boto3==1.2.3
botocore==1.3.26
django-extensions==1.6.1
django-storages==1.4
docutils==0.12
futures==3.0.5
gunicorn==19.4.5
jmespath==0.9.0
psycopg2==2.6.1
python-dateutil==2.4.2
pytz==2015.7
setproctitle==1.1.9
six==1.10.0
wsgiref==0.1.2
